backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
nltk
python-dotenv
requests
